requirements.txt
aniso8601==6.0.0
appdirs==1.4.3
APScheduler==3.6.0
attrs==19.1.0
blinker==1.4
CairoSVG==2.4.2
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
configparser==3.7.4
cssselect==1.0.3
cssselect2==0.2.2
cycler==0.10.0
defusedxml==0.6.0
dnspython==1.16.0
Flask==1.0.2
beautifulsoup4==4.8.2
soupsieve==1.9.5
Flask-Cors==3.0.7
Flask-JWT-Extended==3.18.2
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.7
Flask-Slack==0.1.5
Flask-WeasyPrint==0.6
future==0.17.1
googleapis-common-protos==1.5.10
grpcio==1.22.0
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
isodate==0.6.0
isort==4.3.20
itsdangerous==1.1.0
IMAPClient==2.1.0
six==1.14.0
Jinja2==2.10.1
jsonschema==3.0.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
#lxml==4.3.3
MarkupSafe==1.1.1
pymongo-inmemory==0.1.4
mccabe==0.6.1
mock==3.0.5
numpy==1.16.3
passlib==1.7.1
pdfkit==0.6.1
pdfrw==0.4
Pillow==6.2.0
protobuf==3.7.1
pycparser==2.19
pyfcm==1.4.7
pytest==5.4.3
pytest-flask==1.0.0
PyJWT==1.7.1
pymongo==3.8.0
pyparsing==2.4.0
PyPDF2==1.26.0
Pyphen==0.9.5
pyquery==1.4.0
pyrsistent==0.15.4
PySocks==1.7.0
python-dateutil==2.7.3
python-dotenv==0.10.3
python-http-client==3.2.4
python-pdf==0.37
pytz==2018.7
pyzmail36==1.0.4
reportlab==3.5.26
requests==2.21.0
sendgrid==6.1.1
six==1.14.0
slackclient==1.3.1
soupsieve==1.9.5
tinycss2==1.0.2
tzlocal==2.0.0
urllib3==1.24.2
virtualenv==16.4.1
WeasyPrint==49
twilio==6.39.0 
webencodings==0.5.1
Werkzeug==0.15.5
wkhtmltopdf==0.2
xhtml2pdf==0.2.3
yapf==0.28.0
pyzmail36==1.0.4
six==1.14.0
